pios: ws2811: rename header files
